+++ common/Rhyolite/Schema.hs
@@ -56,13 +56,6 @@ type Email = Text --TODO: Validation
 newtype Json a = Json { unJson :: a }
   deriving (Eq, Ord, Show, ToJSON, FromJSON)
 
--- | Newtype for referring to database large objects. This generally shouldn't have to go over the wire
--- but I'm putting it here where it can be placed in types in the common schema, because often the Ids of
--- those types will want to be shared with the frontend. We're using Word64 here rather than CUInt, which
--- is the type that Oid wraps, because Word64 has Groundhog instances to steal.
-newtype LargeObjectId = LargeObjectId Word64
-  deriving (Eq, Ord, Show, Read, Typeable)
-
 idPathSegmentEncoder
   :: forall a check parse.
   (MonadError Text parse, Applicative check, Show (IdData a), Read (IdData a))